Updated website content
Updated the website's content to include a nice photo underneath the top heading and a helpful little button to quickly copy the email address to the clipboard if preferred over the mailto: link underneath the blurb. Be advised! This copy-address-to-clipboard button does not yet work on mobile. This is the first major addition to the website. thebrewergame.com v1.17