Updated styles.css to hide the copy email button for mobile
styles.css has been updated to hide the copy email button for mobile devices. I hope to have the copy email button be functional for them in the future, but in the meantime, they no longer have to look at a dysfunctional button. The button remains unchanged for non-mobile devices. thebrewergame.com v1.21