Added a tagline to the announcement banner & adjusted email button CSS
The announcement banner now features a little tagline above the header for an additional one-line related to the content! Also, the CSS for the email button has been updated to target that button specifically, as opposed to targeting all buttons on the website like it previously did. Just some futureproofing! thebrewergame.com v1.63